+++ tests/Api/ScreenGroupCampaignTest.php
@@ -25,7 +25,7 @@ class ScreenGroupCampaignTest extends AbstractBaseApiTestCase
         $iri = $this->findIriBy(ScreenGroup::class, ['tenant' => $this->tenant]);
         $screenGroupUlid2 = $this->iriHelperUtils->getUlidFromIRI($iri);
 
-        $client->request('PUT', '/v1/screen-groups/'.$playlistUlid.'/campaigns', [
+        $client->request('PUT', '/v2/screen-groups/'.$playlistUlid.'/campaigns', [
             'json' => [
                 (object) [
                   'screengroup' => $screenGroupUlid1,
@@ -58,13 +58,13 @@ class ScreenGroupCampaignTest extends AbstractBaseApiTestCase
         $iri = $this->findIriBy(ScreenGroup::class, ['tenant' => $this->tenant]);
         $ulid = $this->iriHelperUtils->getUlidFromIRI($iri);
 
-        $client->request('GET', '/v1/screen-groups/'.$ulid.'/campaigns?page=1&itemsPerPage=10', ['headers' => ['Content-Type' => 'application/ld+json']]);
+        $client->request('GET', '/v2/screen-groups/'.$ulid.'/campaigns?page=1&itemsPerPage=10', ['headers' => ['Content-Type' => 'application/ld+json']]);
 
         $this->assertResponseIsSuccessful();
         $this->assertResponseHeaderSame('content-type', 'application/ld+json; charset=utf-8');
         $this->assertJsonContains([
             '@context' => '/contexts/ScreenGroupCampaign',
-            '@id' => '/v1/screen-groups/'.$ulid.'/campaigns',
+            '@id' => '/v2/screen-groups/'.$ulid.'/campaigns',
             '@type' => 'hydra:Collection',
         ]);
     }
@@ -81,7 +81,7 @@ class ScreenGroupCampaignTest extends AbstractBaseApiTestCase
         $iri = $this->findIriBy(ScreenGroup::class, ['tenant' => $this->tenant]);
         $screenGroupUlid2 = $this->iriHelperUtils->getUlidFromIRI($iri);
 
-        $client->request('PUT', '/v1/screen-groups/'.$playlistUlid.'/campaigns', [
+        $client->request('PUT', '/v2/screen-groups/'.$playlistUlid.'/campaigns', [
             'json' => [
                 (object) [
                   'screengroup' => $screenGroupUlid1,
@@ -96,13 +96,13 @@ class ScreenGroupCampaignTest extends AbstractBaseApiTestCase
         ]);
         $this->assertResponseStatusCodeSame(201);
 
-        $client->request('DELETE', '/v1/screen-groups/'.$screenGroupUlid1.'/campaigns/'.$playlistUlid, [
+        $client->request('DELETE', '/v2/screen-groups/'.$screenGroupUlid1.'/campaigns/'.$playlistUlid, [
             'headers' => [
                 'Content-Type' => 'application/ld+json',
             ],
         ]);
         $this->assertResponseStatusCodeSame(204);
-        $client->request('DELETE', '/v1/screen-groups/'.$screenGroupUlid2.'/campaigns/'.$playlistUlid, [
+        $client->request('DELETE', '/v2/screen-groups/'.$screenGroupUlid2.'/campaigns/'.$playlistUlid, [
             'headers' => [
                 'Content-Type' => 'application/ld+json',
             ],